Move errors in a dedicated module
